sw/user/liblitepcie: Only write as many buffers as written by user
We can't send eveything to the driver, we must only send as much
as the user wrote. Also, in case the kernel wasn't ready to accept
all of the data, we need to keep those around. The handling of that
last case is suboptimal ATM (with memmove) but doesn't happen often
in my current tests ...

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>